# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Course {
  id: number;
  name: string;
  code: string;
  color: string;
  /** @nullable */
  instructor?: string | null;
  /** @nullable */
  term?: string | null;
  createdAt: string;
}

export interface CourseInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  color: string;
  instructor?: string;
  term?: string;
}

export interface CourseUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  code?: string;
  color?: string;
  /** @nullable */
  instructor?: string | null;
  /** @nullable */
  term?: string | null;
}

export type TaskPriority = typeof TaskPriority[keyof typeof TaskPriority];


export const TaskPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type TaskType = typeof TaskType[keyof typeof TaskType];


export const TaskType = {
  assignment: 'assignment',
  reading: 'reading',
  exam: 'exam',
  project: 'project',
  other: 'other',
} as const;

export interface Task {
  id: number;
  courseId: number;
  title: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority: TaskPriority;
  type: TaskType;
  completed: boolean;
  createdAt: string;
  updatedAt: string;
}

export type TaskInputPriority = typeof TaskInputPriority[keyof typeof TaskInputPriority];


export const TaskInputPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type TaskInputType = typeof TaskInputType[keyof typeof TaskInputType];


export const TaskInputType = {
  assignment: 'assignment',
  reading: 'reading',
  exam: 'exam',
  project: 'project',
  other: 'other',
} as const;

export interface TaskInput {
  courseId: number;
  /** @minLength 1 */
  title: string;
  notes?: string;
  /** @nullable */
  dueDate?: string | null;
  priority?: TaskInputPriority;
  type?: TaskInputType;
}

export type TaskUpdatePriority = typeof TaskUpdatePriority[keyof typeof TaskUpdatePriority];


export const TaskUpdatePriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type TaskUpdateType = typeof TaskUpdateType[keyof typeof TaskUpdateType];


export const TaskUpdateType = {
  assignment: 'assignment',
  reading: 'reading',
  exam: 'exam',
  project: 'project',
  other: 'other',
} as const;

export interface TaskUpdate {
  courseId?: number;
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority?: TaskUpdatePriority;
  type?: TaskUpdateType;
  completed?: boolean;
}

export interface CourseStat {
  courseId: number;
  name: string;
  code: string;
  color: string;
  totalTasks: number;
  activeTasks: number;
  completedTasks: number;
}

export interface DashboardSummary {
  totalTasks: number;
  activeTasks: number;
  completedTasks: number;
  overdueTasks: number;
  dueSoonTasks: number;
  totalCourses: number;
  courseStats: CourseStat[];
}

export type ListTasksParams = {
courseId?: number;
/**
 * Filter by completion status
 */
status?: ListTasksStatus;
priority?: ListTasksPriority;
type?: ListTasksType;
};

export type ListTasksStatus = typeof ListTasksStatus[keyof typeof ListTasksStatus];


export const ListTasksStatus = {
  all: 'all',
  active: 'active',
  completed: 'completed',
} as const;

export type ListTasksPriority = typeof ListTasksPriority[keyof typeof ListTasksPriority];


export const ListTasksPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export type ListTasksType = typeof ListTasksType[keyof typeof ListTasksType];


export const ListTasksType = {
  assignment: 'assignment',
  reading: 'reading',
  exam: 'exam',
  project: 'project',
  other: 'other',
} as const;

export type GetUpcomingTasksParams = {
/**
 * Number of days ahead to look
 */
days?: number;
};