# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  name: string;
  code: string;
  color: string;
  /** @nullable */
  instructor?: string | null;
  /** @nullable */
  term?: string | null;
  createdAt: string;
}