# lib/api-zod/src/generated/types/courseStat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseStat {
  courseId: number;
  name: string;
  code: string;
  color: string;
  totalTasks: number;
  activeTasks: number;
  completedTasks: number;
}