# lib/api-zod/src/generated/types/courseUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CourseUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  code?: string;
  color?: string;
  /** @nullable */
  instructor?: string | null;
  /** @nullable */
  term?: string | null;
}