# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from './taskPriority';
import type { TaskType } from './taskType';

export interface Task {
  id: number;
  courseId: number;
  title: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority: TaskPriority;
  type: TaskType;
  completed: boolean;
  createdAt: string;
  updatedAt: string;
}