# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';
import type { TaskInputType } from './taskInputType';

export interface TaskInput {
  courseId: number;
  /** @minLength 1 */
  title: string;
  notes?: string;
  /** @nullable */
  dueDate?: string | null;
  priority?: TaskInputPriority;
  type?: TaskInputType;
}