# lib/api-zod/src/generated/types/taskUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskUpdatePriority } from './taskUpdatePriority';
import type { TaskUpdateType } from './taskUpdateType';

export interface TaskUpdate {
  courseId?: number;
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority?: TaskUpdatePriority;
  type?: TaskUpdateType;
  completed?: boolean;
}